requirements.txt
celery==4.3.0
Django==2.2.6
django-admin-sortable2==0.7.3
django-polymorphic==2.1.2
djangorestframework==3.10.3
ipython==7.8.0
psycopg2==2.8.3
pytest==5.2.1
pytest-django==3.5.1
PyYAML==5.1.2